Fix narrowing conversions in album cover fetcher

+++ src/covermanager/albumcoverfetchersearch.cpp
@@ -75,7 +75,7 @@ AlbumCoverFetcherSearch::~AlbumCoverFetcherSearch() {
 void AlbumCoverFetcherSearch::TerminateSearch() {
 
   QList<int> ids = pending_requests_.keys();
-  for (const quint64 id : ids) {
+  for (const int id : ids) {
     pending_requests_.take(id)->CancelSearch(id);
   }
 
@@ -380,7 +380,7 @@ float AlbumCoverFetcherSearch::ScoreImage(const QSize size) {
   const float size_score = std::sqrt(float(size.width() * size.height())) / kTargetSize;
 
   // A 1:1 image scores 1.0, anything else scores less
-  const float aspect_score = float(1.0) - float(std::max(size.width(), size.height()) - std::min(size.width(), size.height())) / std::max(size.height(), size.width());
+  const float aspect_score = float(1.0) - float(std::max(size.width(), size.height()) - std::min(size.width(), size.height())) / float(std::max(size.height(), size.width()));
 
   return size_score + aspect_score;
